Front end (#8)
* Successfully nested settings stack in the profile page

* Added the Payments option page

* Changes made to te payments methods page

* More Changes to Payments Page

* Still Working on To Map Screen Navigation

* chore: Update styling in EPayPage, ePayCard, and HomeScreen

This commit updates the styling in EPayPage, ePayCard, and HomeScreen components. In EPayPage, the background color is changed to '#fafafa' and the flex property is set to 1. In ePayCard, a border is added with a width of 1 and a color of 'rgba(255, 255, 255, 0.3)'. In HomeScreen, the text alignment and margin of the 'Ping the Nearest Service Provider' title are adjusted. The border width of the Fuel and Mechanic grid items is increased to 2, and the border color is changed to '#efa400' and '#0a99f5' respectively. The border width of the Tow Service and Schedule Service grid items is also increased to 2, and the border color is changed to black and '#fc3c42' respectively. The Emergency Contacts section is added with a border, padding, and background color. The version number text is updated to 'Arduiva Beta 0.0.1'.